+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 3.0.1
 * `FIX` [#1033](https://github.com/sumneko/lua-language-server/issues/1033)
+* `FIX` [#1034](https://github.com/sumneko/lua-language-server/issues/1034)
 
 ## 3.0.0
 `2022-4-10`