+++ changelog.md
@@ -4,6 +4,7 @@
 * `FIX` [#1033](https://github.com/sumneko/lua-language-server/issues/1033)
 * `FIX` [#1034](https://github.com/sumneko/lua-language-server/issues/1034)
 * `FIX` [#1035](https://github.com/sumneko/lua-language-server/issues/1035)
+* `FIX` [#1036](https://github.com/sumneko/lua-language-server/issues/1036)
 
 ## 3.0.0
 `2022-4-10`