+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 3.0.1
+`2022-4-11`
 * `FIX` [#1033](https://github.com/sumneko/lua-language-server/issues/1033)
 * `FIX` [#1034](https://github.com/sumneko/lua-language-server/issues/1034)
 * `FIX` [#1035](https://github.com/sumneko/lua-language-server/issues/1035)